Close the twelve findings from the eighth /code-review pass
The two that reach stored data:

A `Date` bound for a TypeORM `date` COLUMN must be LOCAL midnight, not
UTC. `DateUtils.mixedDateToDateString` serializes with local getters, so
`mortgageTermEndDate`'s UTC-midnight return -- and the
`paymentStartDate: new Date(dateOnlyString)` at three creation sites --
were written to PostgreSQL one day early anywhere west of Greenwich. The
computed date was right and the STORED one wrong, which is exactly what
the previous pass's assertions could not see: they went through
`formatDateYMD`, which reads UTC. `localDateForColumn` is the one door,
the timezone spec now asserts through TypeORM's own serializer in child
processes, and a scan fails a bare `new Date(...)` on either property.
The two conventions are opposite and the destination decides which
applies; both say so where they are defined.

Migration 165 healed the frequency and left `next_due_date` frozen where
the broken schedule stalled -- for a two-year-old mortgage, two years in
the past. Unfreezing the cadence against that stale date makes the
auto-post cron drip ONE back-dated installment per run, forty-eight of
them, each moving the balance. The same statement now moves
`next_due_date` to the first date the engine's semi-monthly cadence
reaches on or after today. The occurrences that never fired are
deliberately not posted: inventing forty-eight transactions to answer a
question about somebody's history is the one outcome that cannot be
undone. Verified against PostgreSQL 16, including that a second pass
selects nothing.

`loan-schedule.ts` had grown to 1384 lines against the repository's
800-line ceiling, so it is four modules now -- `loan-frequency.ts`
(cadences, period counts, periodic rates, `advanceDate`),
`loan-overpayments.ts` (plan shapes, `effectiveOverpaymentMode`, the
occurrence calendar), `loan-schedule-types.ts` (the shapes both the
generator and the comparison need, at the bottom of the graph so the
split is acyclic) and `loan-comparison.ts` (rate timeline,
`compareSchedules`) -- with `loan-schedule.ts` re-exporting all of them,
because splitting a file is not a reason to touch fifty call sites. The
largest is 566 lines.

The rest:

- The import wizard's payment-setup dialog omitted the mortgage flags, so
  running setup there on a matched Canadian mortgage turned semi-annual
  compounding off. Same fix as the account form, from the accounts list.
- The rate-change recalculation's new quarterly/yearly fallback dropped
  Canadian compounding, deriving the persisted installment on one
  convention and the split 200 lines above it on another. It keeps the
  account's own compounding; the `calculatePaymentForTermAtPeriods`
  export added for it is gone with its only caller.
- `generateBudgetSchedule` ignored `lowerEndPeriod` and regenerated a
  whole baseline, so the two entry points derived one quantity two ways
  again.
- The PDF export dropped its subtitle entirely on a truncated
  projection, losing the payments-made count with the interest figure.
  It is relabelled instead, as the card beside it already was.
- `t(FREQUENCY_LABEL_KEY[freq])` threw on a cadence the unconstrained
  VARCHAR can hold, taking the loan page down where the engine survives
  the same row. `frequencyLabelKey` degrades.
- The cross-package guard compared which cadences each side accepts, not
  what they map to -- a browser copy returning BIWEEKLY for SEMIMONTHLY
  passed. It compares the values now.
- `roundUpTo`/`minimizeToStep` accumulated float error, so a fractional
  step produced an amount that was not a multiple of it.
- `PAYMENT_FREQUENCIES` and its union were declared in two files held
  together by a runtime spec; the lists moved beside the tables keyed by
  them, so a new cadence is a compile error.
- A JSDoc was orphaned onto the constant above the function it described.
- A labels test pinned a `ScenarioComparison` shape `compareSchedules`
  cannot produce, exercising only the heuristic the mode parameter
  retired.
