+++ frontend/src/components/accounts/loan-detail/loan-scenario-labels.test.ts
@@ -83,18 +83,41 @@ describe('createScenarioLabels', () => {
     );
   });
 
-  it('still reports a lower installment when the schedules cannot be compared', () => {
-    // installmentReduction is a property of each schedule on its own, so the
-    // lower-installment headline survives a truncated comparison.
-    const lowered = {
+  it('says nothing rather than 0.00 when a truncated comparison is lower-installment', () => {
+    // The state this replaced could not occur: it set installmentReduction to a
+    // number beside three nulls, but `compareSchedules` gates all four on the
+    // single `baseline.paidOff && scenario.paidOff` flag, so the reduction is
+    // null exactly when the others are. A fixture the producer cannot produce
+    // pins a branch nothing reaches, and it would have stayed green while the
+    // real path broke.
+    //
+    // The reachable truncated state is all four null, and it is read with the
+    // MODE the caller now always supplies -- the parameter that exists because
+    // inferring lower-installment from a null reduction is what the old
+    // heuristic did.
+    const truncated = {
       scenario: { payoffDate: null, finalPaymentAmount: 400 },
       paymentsSaved: null,
       monthsSaved: null,
       interestSaved: null,
+      installmentReduction: null,
+    } as unknown as ScenarioComparison;
+
+    expect(labels.timeSavedLabel(truncated, 'LOWER_INSTALLMENT')).toBe('—');
+  });
+
+  it('reports the drop when a lower-installment comparison did complete', () => {
+    // The state that IS reachable: both schedules paid off, so all four figures
+    // are numbers, and the mode comes from the plan rather than from the drop.
+    const lowered = {
+      scenario: { payoffDate: '2036-01-15', finalPaymentAmount: 400 },
+      paymentsSaved: 0,
+      monthsSaved: 0,
+      interestSaved: 4200,
       installmentReduction: 100,
     } as unknown as ScenarioComparison;
 
-    expect(labels.timeSavedLabel(lowered)).toBe(
+    expect(labels.timeSavedLabel(lowered, 'LOWER_INSTALLMENT')).toBe(
       'loanDetail.comparison.installmentDrop:$400.00,$100.00',
     );
   });
