+++ frontend/src/i18n/messages/es/reports.json
@@ -842,7 +842,8 @@
     "typeLineOfCredit": "Línea de crédito",
     "typeLoan": "Préstamo",
     "typeMortgage": "Hipoteca",
-    "pdfSubtitlePaymentsSummary": "{count} pagos realizados, {interest} en intereses totales"
+    "pdfSubtitlePaymentsSummary": "{count} pagos realizados, {interest} en intereses totales",
+    "pdfSubtitlePaymentsProjection": "{count} pagos realizados, {interest} de intereses durante la proyección"
   },
   "monteCarlo": {
     "toasts": {
