+++ frontend/src/i18n/messages/ja/reports.json
@@ -842,7 +842,8 @@
     "typeLineOfCredit": "当座貸越",
     "typeLoan": "ローン",
     "typeMortgage": "住宅ローン",
-    "pdfSubtitlePaymentsSummary": "{count} 件の支払い済み、{interest} 総利息"
+    "pdfSubtitlePaymentsSummary": "{count} 件の支払い済み、{interest} 総利息",
+    "pdfSubtitlePaymentsProjection": "{count} 回支払済み、予測期間の利息 {interest}"
   },
   "monteCarlo": {
     "toasts": {
