+++ frontend/src/i18n/messages/ko/reports.json
@@ -842,7 +842,8 @@
     "typeLineOfCredit": "신용 한도",
     "typeLoan": "대출",
     "typeMortgage": "주택 담보 대출",
-    "pdfSubtitlePaymentsSummary": "{count}건의 납부 완료, 총 이자 {interest}"
+    "pdfSubtitlePaymentsSummary": "{count}건의 납부 완료, 총 이자 {interest}",
+    "pdfSubtitlePaymentsProjection": "{count}회 상환 완료, 예상 기간 이자 {interest}"
   },
   "monteCarlo": {
     "toasts": {
