+++ frontend/src/i18n/messages/vi/reports.json
@@ -842,7 +842,8 @@
     "typeLineOfCredit": "Hạn mức tín dụng",
     "typeLoan": "Vay",
     "typeMortgage": "Thế chấp",
-    "pdfSubtitlePaymentsSummary": "Đã thực hiện {count} lần thanh toán, tổng lãi {interest}"
+    "pdfSubtitlePaymentsSummary": "Đã thực hiện {count} lần thanh toán, tổng lãi {interest}",
+    "pdfSubtitlePaymentsProjection": "Đã thanh toán {count} lần, lãi {interest} trong kỳ dự phóng"
   },
   "monteCarlo": {
     "toasts": {
