+++ frontend/src/lib/loan-schedule.ts
@@ -1,7 +1,3 @@
-import { parseLocalDate } from '@/lib/utils';
-import { advanceByFrequency } from '@/lib/frequency';
-import type { FrequencyType } from '@/types/scheduled-transaction';
-
 /**
  * Shared loan/mortgage schedule engine.
  *
@@ -21,770 +17,36 @@ import type { FrequencyType } from '@/types/scheduled-transaction';
  */
 
 /**
- * Payment frequencies a schedule can be projected at.
- *
- * Semi-monthly appears twice on purpose: `accounts.payment_frequency` holds
- * whichever spelling wrote it. The mortgage path stores the mortgage enum's
- * `SEMI_MONTHLY`; the loan-payment setup dialog stores the scheduled-transaction
- * recurrence's `SEMIMONTHLY`. Accepting only the first made
- * `buildLoanProjectionInput`'s cast fall through `getPeriodsPerYear` to its
- * monthly default, so a semi-monthly loan was projected at 12 periods a year
- * instead of 24 -- roughly double the remaining interest and a payoff date twice
- * as far out, on every surface that reads a projection.
- * `frontend/src/lib/loan-frequency.guard.test.ts` holds both spellings.
+ * The schedule engine. Cadences live in `loan-frequency.ts` and overpayment
+ * plans in `loan-overpayments.ts`; both are re-exported here because every
+ * consumer has always imported from this module, and splitting a file is not a
+ * reason to touch fifty call sites.
  */
-export type ScheduleFrequency =
-  | 'WEEKLY'
-  | 'BIWEEKLY'
-  | 'SEMI_MONTHLY'
-  | 'SEMIMONTHLY'
-  | 'MONTHLY'
-  | 'QUARTERLY'
-  | 'YEARLY'
-  | 'ACCELERATED_WEEKLY'
-  | 'ACCELERATED_BIWEEKLY';
+export * from "@/lib/loan-frequency";
+export * from "@/lib/loan-overpayments";
+export * from "@/lib/loan-schedule-types";
+export * from "@/lib/loan-comparison";
 
-/**
- * A local calendar day as `yyyy-MM-dd`.
- *
- * date-fns `format` with this pattern, without the pattern parser: the schedule
- * loop calls it once per row, and a goal-seek runs thirty schedules of up to
- * `HARD_MAX_PAYMENTS` rows per keystroke, so it is the hottest line in the
- * engine. Local components, matching `format`'s reading of a local `Date` and
- * `parseLocalDate`'s construction of one.
- *
- * The invalid-date throw is kept deliberately: `format` raises `RangeError` on
- * one, and a hand-rolled builder that emits "NaN-NaN-NaN" instead would put that
- * string in a row date, a payoff date and a saved scenario rather than failing.
- */
-function isoDay(date: Date): string {
-  if (Number.isNaN(date.getTime())) {
-    throw new RangeError('Invalid time value');
-  }
-  const month = date.getMonth() + 1;
-  const day = date.getDate();
-  return `${date.getFullYear()}-${month < 10 ? '0' : ''}${month}-${day < 10 ? '0' : ''}${day}`;
-}
-
-export interface LumpSum {
-  /** ISO date (yyyy-MM-dd) the lump sum is paid */
-  date: string;
-  amount: number;
-  /**
-   * Whether this overpayment shortens the term (keep the installment) or lowers
-   * the installment (keep the end date). Defaults to SHORTEN_TERM when omitted.
-   */
-  mode?: OverpaymentMode;
-}
-
-/**
- * How often a recurring overpayment is made. ONE_OFF is a single dated payment
- * (modelled as a lump sum, not a RecurringExtra); the rest recur.
- *
- * A cadence is a *calendar*, independent of the loan's own payment frequency:
- * occurrences fall on the anchor date and every cadence step after it, and each
- * is applied at the first loan payment on or after its due date
- * (`recurringOccurrencesDue`). So MONTHLY contributes exactly 12 occurrences per
- * calendar year on a weekly, biweekly or monthly loan, and the borrower's
- * nominal annual cash is what they typed. Deriving a payment interval instead
- * (`round(periodsPerYear / overpaymentsPerYear)`) paid a monthly extra 13 times
- * a year on a biweekly loan.
- */
-export type OverpaymentFrequency =
-  | 'ONE_OFF'
-  | 'WEEKLY'
-  | 'BIWEEKLY'
-  | 'MONTHLY'
-  | 'QUARTERLY'
-  | 'ANNUALLY';
-
-/**
- * The frequencies that genuinely recur. `ONE_OFF` is a single dated payment,
- * modelled as a `LumpSum`, so it is not a cadence a `RecurringExtra` can carry.
- */
-export type RecurringOverpaymentFrequency = Exclude<
-  OverpaymentFrequency,
-  'ONE_OFF'
->;
-
-/**
- * Overpayments per year for each frequency. `ONE_OFF` is 0 -- a single dated
- * payment has no annual rate, and a caller spreading an amount over a year must
- * treat it separately rather than divide by zero occurrences.
- *
- * A `Record` rather than a `switch`, so a new member of `OverpaymentFrequency`
- * is a compile error here instead of a silent `default: 0`.
- */
-const OVERPAYMENTS_PER_YEAR: Record<OverpaymentFrequency, number> = {
-  ONE_OFF: 0,
-  WEEKLY: 52,
-  BIWEEKLY: 26,
-  MONTHLY: 12,
-  QUARTERLY: 4,
-  ANNUALLY: 1,
-};
-
-/** Overpayments per year for each recurring frequency (ONE_OFF is not recurring). */
-export function overpaymentsPerYear(frequency: OverpaymentFrequency): number {
-  return OVERPAYMENTS_PER_YEAR[frequency] ?? 0;
-}
-
-/**
- * The *average* extra per loan payment for a recurring overpayment of `amount`
- * per `frequency`: the nominal annual cash spread over the loan's periods.
- *
- * Display only -- it is what the "resulting monthly payment" card adds to the
- * installment when the cadence is at least as frequent as the loan's payments.
- * The schedule engine does NOT apply this figure; it dates each occurrence and
- * applies the full amount (see `recurringOccurrencesDue`), so a balance must
- * never be computed from this average.
- */
-export function perPaymentExtraAmount(
-  amount: number,
-  frequency: OverpaymentFrequency,
-  loanFrequency: ScheduleFrequency,
-): number {
-  const per = overpaymentsPerYear(frequency);
-  if (per <= 0) return 0;
-  return (amount * per) / getPeriodsPerYear(loanFrequency);
-}
-
-export interface RecurringExtra {
-  /** Amount paid per occurrence of `frequency` (not per loan payment). */
-  amount: number;
-  /** ISO date (yyyy-MM-dd); applies from the first payment when omitted */
-  startDate?: string;
-  /** ISO date (yyyy-MM-dd); applies until payoff when omitted */
-  endDate?: string;
-  /**
-   * Cadence of the overpayment. Omitted means the amount is applied on every
-   * loan payment as-is (legacy "extra per payment"); a set frequency dates the
-   * occurrences on the calendar and applies the full amount at the first loan
-   * payment on or after each due date.
-   *
-   * `ONE_OFF` is excluded by the type rather than by convention: it is a single
-   * dated payment, so it belongs in `lumpSums`. Accepted here it collapsed into
-   * the legacy no-cadence branch and paid the full amount on EVERY payment --
-   * a lump sum of 5000 becoming 60,000 a year. The backend enum does not admit
-   * it, but nothing in the types said so.
-   */
-  frequency?: RecurringOverpaymentFrequency;
-  /**
-   * Whether this overpayment shortens the term or lowers the installment.
-   * Defaults to SHORTEN_TERM when omitted.
-   */
-  mode?: OverpaymentMode;
-}
-
-export interface OverpaymentPlan {
-  recurringExtra?: RecurringExtra;
-  lumpSums?: LumpSum[];
-  /**
-   * A fixed total to spend on the loan each period (installment + overpayment).
-   * Modelled in the lower-installment style: every period the installment is
-   * recomputed over the remaining contractual term, and the rest of the budget
-   * is overpaid -- so as the installment falls the overpayment grows to keep the
-   * total constant. When set, recurringExtra and lumpSums are ignored.
-   */
-  targetMonthlyPayment?: number;
-  /**
-   * How the budget's installment/overpayment split is shown. LOWER_INSTALLMENT
-   * (default, matching how banks apply overpayments) re-amortizes the
-   * installment each period (it shrinks, the overpayment grows); SHORTEN_TERM
-   * keeps the contractual installment fixed and the overpayment constant. The
-   * balance and payoff are identical either way -- only the split differs.
-   */
-  targetMonthlyPaymentMode?: OverpaymentMode;
-  /** ISO date (yyyy-MM-dd); the budget applies from the first payment when
-   *  omitted. Before it, only the regular installment is paid. */
-  targetMonthlyPaymentStart?: string;
-  /** ISO date (yyyy-MM-dd); the budget applies until payoff when omitted. After
-   *  it, the loan reverts to the regular installment. */
-  targetMonthlyPaymentEnd?: string;
-}
-
-/**
- * The mode a plan actually behaves as, from whichever of its three carriers is
- * in force.
- *
- * A plan can name its mode in three places -- `targetMonthlyPaymentMode`,
- * `recurringExtra.mode`, and each lump sum's -- and reading only the second was
- * how the heuristic it replaced kept coming back: a saved BUDGET scenario has
- * `recurringExtraMode = null` and its mode on `targetMonthlyPaymentMode`, so the
- * caller fell through to `installmentReduction > 0.005`, which is null exactly
- * when a schedule truncated -- the case the explicit mode was added for.
- *
- * The precedence follows the engine, not a preference. A budget IGNORES
- * `recurringExtra` and `lumpSums` (see `OverpaymentPlan.targetMonthlyPayment`),
- * so when one is set its mode is the only one that describes anything, and its
- * default matches the engine's `?? 'LOWER_INSTALLMENT'`. Otherwise the engine
- * re-levels the installment when ANY overpayment asks it to
- * (`anyLowerOverpayment`), so any LOWER_INSTALLMENT carrier makes the plan one:
- * a consumer that adds the overpayment on top of an installment a lump sum has
- * already re-levelled down counts the same money twice.
- *
- * `null` means the plan carries no overpayment at all -- distinct from
- * SHORTEN_TERM, which is a plan that overpays and keeps its installment.
- */
-export function effectiveOverpaymentMode(
-  plan: OverpaymentPlan | undefined | null,
-): OverpaymentMode | null {
-  if (!plan) return null;
-  if ((plan.targetMonthlyPayment ?? 0) > 0) {
-    return plan.targetMonthlyPaymentMode ?? 'LOWER_INSTALLMENT';
-  }
-  const carriers = [
-    ...(plan.recurringExtra ? [plan.recurringExtra.mode] : []),
-    ...(plan.lumpSums ?? []).map((lump) => lump.mode),
-  ];
-  if (carriers.length === 0) return null;
-  return carriers.some((mode) => (mode ?? 'SHORTEN_TERM') === 'LOWER_INSTALLMENT')
-    ? 'LOWER_INSTALLMENT'
-    : 'SHORTEN_TERM';
-}
-
-/**
- * What a bank holds fixed after an overpayment:
- * - SHORTEN_TERM (PL *skrócenie okresu*): keep the installment, pay off sooner.
- * - LOWER_INSTALLMENT (PL *obniżenie raty*): keep the end date, recompute a
- *   smaller installment that amortizes the reduced balance over the remaining
- *   periods.
- */
-export type OverpaymentMode = 'SHORTEN_TERM' | 'LOWER_INSTALLMENT';
-
-/** A step on the loan's interest-rate timeline, applied during generation */
-export interface RateChange {
-  /** ISO date (yyyy-MM-dd) the new rate takes effect */
-  effectiveDate: string;
-  /** Annual rate as a percentage, e.g. 4.9 */
-  annualRate: number;
-  /** New regular payment from this date; omitted/null = payment unchanged */
-  paymentAmount?: number | null;
-}
-
-/** A persisted rate-history row, as returned by the rate-changes API */
-export interface RateTimelineRow {
-  effectiveDate: string;
-  annualRate: number;
-  newPaymentAmount?: number | null;
-}
-
-export interface RateTimeline {
-  /** Rate in effect at the schedule start */
-  startingAnnualRate: number;
-  /** Payment in effect at the schedule start, when the timeline knows it */
-  startingPaymentAmount: number | null;
-  /** Steps dated after the schedule start, ready for generateLoanSchedule */
-  rateChanges: RateChange[];
-}
-
-export interface LoanScheduleInput {
-  /** Positive remaining balance to amortize */
-  startingBalance: number;
-  /** Annual rate as a percentage, e.g. 5.5 */
-  annualRate: number;
-  /** Regular contractual payment per period */
-  paymentAmount: number;
-  frequency: ScheduleFrequency;
-  /** Canadian fixed-rate mortgages compound semi-annually */
-  isCanadian?: boolean;
-  isVariableRate?: boolean;
-  /** Date of the first projected payment (row 1) */
-  firstPaymentDate: Date;
-  overpayments?: OverpaymentPlan;
-  /** Known rate steps; each applies from the first payment on/after its date */
-  rateChanges?: RateChange[];
-  /**
-   * Maximum projected payments. Defaults to `DEFAULT_MAX_PROJECTION_YEARS`
-   * worth of this frequency's payments (`maxPaymentsForHorizon`), clamped to
-   * `HARD_MAX_PAYMENTS`.
-   */
-  maxPayments?: number;
-  /**
-   * Amortize to zero over exactly this many payments, re-levelling the payment
-   * each period (so it also adjusts on every rate change). Models a
-   * variable-rate loan that holds its term by adjusting the installment when
-   * the rate moves, so a fixed payment can neither stall nor stretch the
-   * schedule. Superseded by the LOWER_INSTALLMENT overpayment mode, which
-   * derives its own fixed end from the baseline.
-   */
-  fixedEndPeriod?: number;
-  /**
-   * Term (in periods) to re-level the installment toward ONLY when a rate rise
-   * would push the current payment below the period's interest (a stall).
-   * Unlike `fixedEndPeriod` it does not re-level every period, so a schedule can
-   * follow its real recorded payment amounts and still be rescued from a stall
-   * instead of stopping. Superseded by `fixedEndPeriod` (which already rescues).
-   */
-  rescueEndPeriod?: number;
-  /**
-   * The no-overpayment payoff length, when the caller has already computed it.
-   *
-   * A LOWER_INSTALLMENT overpayment re-levels the installment toward the term
-   * the loan would have run WITHOUT any overpayment, so the engine derives that
-   * term by generating a whole second schedule with `overpayments: undefined`.
-   * Inside a goal-seek that is the same schedule thirty times over -- the solver
-   * already computes it once as its own baseline, and every candidate differs
-   * only in the overpayment. Supplying it here skips the recursion; the value
-   * MUST be `generateLoanSchedule({ ...input, overpayments: undefined })
-   * .numPayments` for this exact input, or the installment is re-levelled toward
-   * a term the loan does not have.
-   *
-   * Ignored when `fixedEndPeriod` is set, which supersedes it, and when the plan
-   * carries no LOWER_INSTALLMENT overpayment, which needs no such term.
-   */
-  lowerEndPeriod?: number;
-  /** Seed for cumulative principal (e.g. historical principal already paid) */
-  initialCumulativePrincipal?: number;
-  /** Seed for cumulative interest (e.g. historical interest already paid) */
-  initialCumulativeInterest?: number;
-}
-
-export interface ScheduleRow {
-  paymentNumber: number;
-  /** ISO date (yyyy-MM-dd) */
-  date: string;
-  /** Regular payment applied this period (principal + interest) */
-  payment: number;
-  /** Principal portion of the regular payment */
-  principal: number;
-  interest: number;
-  /** Recurring extra + lump sums applied this period */
-  extraPrincipal: number;
-  /** Balance after this payment */
-  balance: number;
-  /** Annual rate (percentage) in effect for this payment */
-  annualRate: number;
-  /** Running principal incl. extra, seeded by initialCumulativePrincipal */
-  cumulativePrincipal: number;
-  /** Running interest, seeded by initialCumulativeInterest */
-  cumulativeInterest: number;
-}
-
-export interface LoanScheduleResult {
-  rows: ScheduleRow[];
-  /** Date of the final payment, or null when not paid off within maxPayments */
-  payoffDate: string | null;
-  /**
-   * Interest over the rows actually projected. This is the loan's **lifetime**
-   * interest only when `paidOff` is true; when the schedule stopped at the
-   * projection horizon it is the interest over that horizon -- a subtotal, per
-   * `docs/financial-calculation-contract.md` section 1. Every consumer that
-   * presents a lifetime figure, or a saving derived from one, must gate on
-   * `paidOff` first. Which consumers do is recorded, with an honest status, in
-   * INV-LOAN-002 -- not counted here, where the number would rot.
-   */
-  totalInterest: number;
-  /** Regular payments + extra principal contributed across the schedule */
-  totalPaid: number;
-  totalExtraPrincipal: number;
-  numPayments: number;
-  paidOff: boolean;
-  /**
-   * The regular installment in effect at the end of the schedule. Equal to the
-   * contractual payment for SHORTEN_TERM; the recomputed lower payment for
-   * LOWER_INSTALLMENT (PL *obniżenie raty*).
-   *
-   * "The end of the schedule" is the last row projected, which is the last
-   * PAYMENT only when `paidOff` is true. On a schedule truncated by the horizon
-   * this is a mid-schedule installment, so anything presenting it as final --
-   * `compareSchedules().installmentReduction` is the one that does -- gates on
-   * `paidOff` first.
-   */
-  finalPaymentAmount: number;
-}
-
-export interface ScenarioComparison {
-  baseline: LoanScheduleResult;
-  scenario: LoanScheduleResult;
-  /**
-   * Payments the scenario saves against the baseline, or `null` when either
-   * schedule stopped at the projection horizon instead of paying off -- a
-   * horizon's row count minus a lifetime's is not a number of payments saved.
-   */
-  paymentsSaved: number | null;
-  /**
-   * Months the scenario saves, or `null` on the same condition. A truncated
-   * schedule has no payoff date, and "0 months" is a claim that the overpayment
-   * bought no time rather than that the answer is unknown.
-   */
-  monthsSaved: number | null;
-  /**
-   * Interest the scenario saves against the baseline, or `null` when either
-   * schedule stopped at the projection horizon instead of paying off. The
-   * difference of two horizons is not a saving, and presenting one as though it
-   * were is how a truncated baseline contaminated every scenario's headline.
-   */
-  interestSaved: number | null;
-  /**
-   * How much lower the scenario's ending installment is than the baseline's.
-   * Zero for SHORTEN_TERM (the installment is unchanged); positive for
-   * LOWER_INSTALLMENT (PL *obniżenie raty*); `null` when either schedule stopped
-   * at the projection horizon.
-   *
-   * A truncated schedule has no ending installment: `finalPaymentAmount` is
-   * whatever the re-levelled payment happened to be at the horizon's last row,
-   * mid-schedule. Reporting a drop from it put "New Installment: X (-Y)" beside
-   * "Unknown" for time and interest saved on the same card row.
-   */
-  installmentReduction: number | null;
-}
-
-/**
- * Default projection horizon, in years. The bound on a schedule is a span of
- * time, not a row count: a flat 600-payment default was 50 years of a monthly
- * loan but only 11 years of a weekly one, so ordinary 25- and 30-year weekly and
- * biweekly mortgages (1300 and 780 payments) stopped mid-term and reported no
- * payoff date and a fraction of their lifetime interest.
- *
- * 50 years is past the longest amortization any supported term expresses, so a
- * schedule that still has not cleared is genuinely not amortizing rather than
- * merely long.
- */
-export const DEFAULT_MAX_PROJECTION_YEARS = 50;
-
-/**
- * Absolute ceiling on projected rows, whatever a caller asks for. Exported
- * because `loan-past-impact.ts` deliberately projects to it, and a literal copy
- * there would silently disagree the moment this moves.
- */
-export const HARD_MAX_PAYMENTS = 10000;
-/** Balances at or below this are considered paid off (matches the reports) */
-const PAYOFF_EPSILON = 0.01;
-
-/**
- * Payments in `years` of `frequency`, clamped to `HARD_MAX_PAYMENTS` -- the
- * default projection bound, and the one place the horizon becomes a row count.
- */
-export function maxPaymentsForHorizon(
-  frequency: ScheduleFrequency,
-  years: number = DEFAULT_MAX_PROJECTION_YEARS,
-): number {
-  return Math.min(
-    Math.max(1, Math.round(getPeriodsPerYear(frequency) * years)),
-    HARD_MAX_PAYMENTS,
-  );
-}
-
-/**
- * The row cap a schedule runs to: a supplied `maxPayments` or the frequency's
- * horizon, floored at one row and clamped to the hard maximum.
- *
- * Written once because both entry points need it and had drifted: the budget
- * path omitted the `Math.max(1, ...)` floor, so `maxPayments: 0` gave
- * `generateLoanSchedule` one row and `generateBudgetSchedule` zero -- an empty,
- * not-paid-off result for an input the other path amortized.
- */
-function resolveMaxPayments(
-  frequency: ScheduleFrequency,
-  maxPayments: number | undefined,
-): number {
-  return Math.min(
-    Math.max(1, maxPayments ?? maxPaymentsForHorizon(frequency)),
-    HARD_MAX_PAYMENTS,
-  );
-}
-
-/**
- * Payment periods a year for each cadence a loan account can store.
- *
- * A `Record` rather than a `switch` with a `default: 12`: the default is how
- * semi-monthly was silently projected at twelve periods a year -- roughly double
- * the remaining interest and a payoff date twice as far out -- and a table makes
- * a new member of `ScheduleFrequency` a compile error instead. The runtime
- * fallback below survives only for the cast at `buildLoanProjectionInput`, where
- * a value out of the database has not been through the compiler at all.
- */
-const PERIODS_PER_YEAR: Record<ScheduleFrequency, number> = {
-  WEEKLY: 52,
-  ACCELERATED_WEEKLY: 52,
-  BIWEEKLY: 26,
-  ACCELERATED_BIWEEKLY: 26,
-  SEMI_MONTHLY: 24,
-  SEMIMONTHLY: 24,
-  MONTHLY: 12,
-  QUARTERLY: 4,
-  YEARLY: 1,
-};
-
-export function getPeriodsPerYear(frequency: ScheduleFrequency): number {
-  return PERIODS_PER_YEAR[frequency] ?? 12;
-}
-
-/**
- * Periodic rate, on the convention named in docs/financial-semantics.md
- * section 9: the quoted rate is a nominal annual rate compounded at the payment
- * frequency, so the periodic rate is `annualRate / periodsPerYear` -- NOT
- * monthly compounding converted to the period. Canadian fixed-rate mortgages
- * are the one exception (semi-annual compounding, required by law).
- *
- * Mirrors `getPeriodicRate` in
- * backend/src/accounts/mortgage-amortization.util.ts, which carries the full
- * reasoning. The mirroring detects drift; it is not evidence for the convention.
- */
-export function getPeriodicRate(
-  annualRate: number,
-  periodsPerYear: number,
-  isCanadian: boolean,
-  isVariableRate: boolean,
-): number {
-  if (annualRate === 0) return 0;
-  if (isCanadian && !isVariableRate) {
-    // Canadian fixed-rate: semi-annual compounding
-    const semiAnnualRate = annualRate / 100 / 2;
-    return Math.pow(1 + semiAnnualRate, 2 / periodsPerYear) - 1;
-  }
-  return annualRate / 100 / periodsPerYear;
-}
-
-/**
- * Effective annual rate as a percentage, unrounded -- the rate the loan actually
- * costs over a year, compounded the way its own periodic rate is derived.
- *
- * The one frontend implementation of the "Displayed EAR" row of
- * docs/financial-semantics.md section 9, mirroring the backend's
- * `calculateEffectiveAnnualRate` (which rounds to 2dp for its API contract; this
- * returns the raw percentage so each surface chooses its own precision). The
- * loan detail summary card previously computed the Canadian formula inline,
- * which made it a third copy of the compounding convention that the
- * frequency-aware fix never reached.
- *
- * @param periodsPerYear - Payments per year, from getPeriodsPerYear
- */
-export function effectiveAnnualRate(
-  annualRate: number,
-  periodsPerYear: number,
-  isCanadian: boolean,
-  isVariableRate: boolean,
-): number {
-  if (isCanadian && !isVariableRate) {
-    // Semi-annual compounding, required by law and independent of how often the
-    // mortgage is paid.
-    return (Math.pow(1 + annualRate / 100 / 2, 2) - 1) * 100;
-  }
-  // Nominal rate compounded at the payment frequency.
-  return (Math.pow(1 + annualRate / 100 / periodsPerYear, periodsPerYear) - 1) * 100;
-}
-
-/**
- * Interest accrued on `balance` over one payment period at `annualRate`. A
- * candidate installment amortizes only when it exceeds this, so it is the
- * shared guard for seeding a projection or the contractual schedule -- rejecting
- * a principal-only figure that would never reduce the balance.
- */
-export function firstPeriodInterest(
-  balance: number,
-  annualRate: number,
-  frequency: ScheduleFrequency,
-  isCanadian = false,
-  isVariableRate = false,
-): number {
-  return (
-    balance *
-    getPeriodicRate(annualRate, getPeriodsPerYear(frequency), isCanadian, isVariableRate)
-  );
-}
-
-/**
- * The recurrence frequency each loan cadence is posted at.
- *
- * The projection's row dates and the scheduler's posting dates are the same
- * calendar to a borrower, so they are the same calendar in code: this table maps
- * a loan cadence onto the recurrence engine's own frequency, and `advanceDate`
- * steps through `advanceByFrequency`. The accelerated cadences pay a fraction of
- * the monthly installment on their base cadence, so they step as that base.
- *
- * A `Record`, so a new member of `ScheduleFrequency` is a compile error rather
- * than a silent monthly `default`. It is the browser-side twin of the backend's
- * `LOAN_FREQUENCY_TO_RECURRENCE` / `MORTGAGE_FREQUENCY_TO_RECURRENCE`
- * (`backend/src/accounts/payment-frequency.util.ts`).
- */
-const SCHEDULE_FREQUENCY_TO_RECURRENCE: Record<ScheduleFrequency, FrequencyType> = {
-  WEEKLY: 'WEEKLY',
-  ACCELERATED_WEEKLY: 'WEEKLY',
-  BIWEEKLY: 'BIWEEKLY',
-  ACCELERATED_BIWEEKLY: 'BIWEEKLY',
-  SEMI_MONTHLY: 'SEMIMONTHLY',
-  SEMIMONTHLY: 'SEMIMONTHLY',
-  MONTHLY: 'MONTHLY',
-  QUARTERLY: 'QUARTERLY',
-  YEARLY: 'YEARLY',
-};
-
-/**
- * The next payment date after `date` on the loan's cadence.
- *
- * Delegated to `advanceByFrequency`, the engine that posts these payments, and
- * not a second calendar beside it. The hand-rolled version got two things wrong
- * at once. Semi-monthly stepped the 1st and the 15th where the engine steps the
- * 15th and the last day of the month, showing a borrower dates their register
- * never has. And the month cadences used `Date.setMonth(+1)`, which OVERFLOWS
- * rather than clamps: a loan paid on the 31st had its second row dated 3 March
- * -- February skipped entirely -- and every later row three days off the
- * schedule the backend's `calculateEndDate` bounds. `advanceByFrequency` clamps
- * (`addMonthsClamped`), so 31 January steps to 28 February, and INV-LOAN-005's
- * requirement -- that a projected payoff is a date the scheduler reaches --
- * holds by construction on this side too.
- *
- * `loan-frequency.guard.test.ts` walks every cadence against the engine,
- * including the month-end anchors, so the two cannot drift apart again.
- */
-export function advanceDate(date: Date, frequency: ScheduleFrequency): Date {
-  return advanceByFrequency(date, SCHEDULE_FREQUENCY_TO_RECURRENCE[frequency] ?? 'MONTHLY');
-}
-
-/**
- * The recurrence frequency each overpayment cadence steps on.
- *
- * The same engine the loan's own payment dates use, and for the same reason: an
- * occurrence is applied at the first loan payment on or after its due date, so
- * two calendars that clamp differently make an occurrence miss the payment it
- * belongs to. They did. Occurrences were derived from the anchor by index and
- * clamped per month (31 Jan, 28 Feb, 31 Mar, 30 Apr ...) while the loan's rows
- * accumulate the engine's clamp (31 Jan, 28 Feb, 28 Mar, 28 Apr ...), so on a
- * monthly loan first paid on the 31st the occurrence due 31 March arrived after
- * the row dated 28 March, waited for 28 April, and the borrower paid ELEVEN
- * monthly extras in the first calendar year and two on one row the next
- * February -- the count INV-LOAN-001 exists to protect, broken by the calendars
- * disagreeing rather than by the arithmetic.
- *
- * The price is the one the anchor-derived version was written to avoid: an
- * accumulating clamp is lossy, so a cadence anchored on the 31st settles onto
- * the 28th after its first February instead of returning to month-end the way a
- * standing order would. On a loan paid on the 31st that is not a cost -- the
- * loan's own payments have settled onto the 28th too, and the overpayment
- * follows them -- and for any anchor on the 28th or earlier the two are
- * identical. The count per year, which is the invariant, is preserved either
- * way; the alignment is not.
- */
-const CADENCE_RECURRENCE: Record<
-  RecurringOverpaymentFrequency,
-  FrequencyType
-> = {
-  WEEKLY: 'WEEKLY',
-  BIWEEKLY: 'BIWEEKLY',
-  MONTHLY: 'MONTHLY',
-  QUARTERLY: 'QUARTERLY',
-  ANNUALLY: 'YEARLY',
-};
-
-/** Counts the overpayment occurrences a given loan payment has to carry. */
-export interface RecurringOccurrenceCounter {
-  /**
-   * Occurrences due on or before `rowDate` that no earlier payment has taken.
-   * Stateful: call once per schedule row, in date order.
-   */
-  dueBy(rowDate: string): number;
-}
-
-/**
- * The one place a recurring overpayment's cadence becomes schedule rows.
- *
- * A declared frequency is a **calendar**: occurrences fall on the anchor date
- * and every cadence step after it, and each is applied at the first loan payment
- * on or after its due date. That makes the annual count exact in both
- * directions -- 12 monthly occurrences a year on a weekly, biweekly or monthly
- * loan; 52 weekly occurrences a year on a monthly loan, arriving four or five at
- * a time -- so the borrower's nominal annual cash is what they typed.
- *
- * The interval this replaced was `round(periodsPerYear / overpaymentsPerYear)`,
- * which is only exact when the ratio divides: `26 / 12` rounded to 2, so a
- * "monthly" 100 landed every second biweekly payment, 13 times a year.
- *
- * Two window rules, both deliberate:
- * - The anchor never precedes the first projected payment. A start date already
- *   in the past means "from now", not a backlog of missed occurrences dumped
- *   onto row 1.
- * - The end date is tested against an occurrence's **due** date, not against
- *   the payment that carries it: cash committed inside the window is still
- *   paid, even when the next payment date falls past the end.
- *
- * An omitted frequency keeps the legacy meaning -- the amount on every payment
- * inside the window, so every payment is exactly one occurrence.
- */
-export function recurringOccurrencesDue(
-  extra: RecurringExtra,
-  firstPaymentDate: Date,
-): RecurringOccurrenceCounter {
-  // Three cases, not two. No frequency declared is the legacy "amount on every
-  // payment"; a recognized cadence is dated below; a frequency declared but
-  // unrecognized is neither.
-  //
-  // The third is reachable even though `RecurringExtra.frequency` excludes
-  // ONE_OFF by type: `loan_scenarios.recurring_extra_frequency` is an
-  // unconstrained VARCHAR whose only enforcement is `@IsIn` at write time, so a
-  // legacy row or a restored backup can carry anything. An undefined lookup
-  // landing in the legacy branch applied the FULL amount on every payment --
-  // the densest possible reading of an unknown value, turning a declared 5000
-  // annual overpayment into 5000 a month with a payoff and a saving to match.
-  // Contributing nothing is the direction that cannot invent a saving.
-  const declaredCadence = extra.frequency
-    ? (CADENCE_RECURRENCE[extra.frequency] as FrequencyType | undefined)
-    : null;
-
-  if (declaredCadence === undefined) {
-    return { dueBy: () => 0 };
-  }
-
-  const cadence = declaredCadence;
-  if (!cadence) {
-    return {
-      dueBy: (rowDate) =>
-        (!extra.startDate || extra.startDate <= rowDate) &&
-        (!extra.endDate || rowDate <= extra.endDate)
-          ? 1
-          : 0,
-    };
-  }
-
-  const firstPaymentIso = isoDay(firstPaymentDate);
-  const anchor =
-    extra.startDate && extra.startDate > firstPaymentIso
-      ? parseLocalDate(extra.startDate)
-      : new Date(firstPaymentDate);
-
-  // Occurrence 0 is the anchor itself; each consumed occurrence steps the cursor
-  // once, so the walk is one engine step per occurrence rather than a fresh
-  // derivation per row -- and consuming in order is already this counter's
-  // contract (the out-of-order guard below).
-  let cursor = new Date(anchor);
-  let lastRowDate: string | null = null;
-  // The pending occurrence's ISO date, cached because most rows consume nothing:
-  // re-deriving and re-formatting it per row cost one date-fns `format` on every
-  // one of a weekly loan's 2600 projected payments, and the goal-seek solver
-  // builds ~24 schedules per keystroke.
-  let pendingDueIso: string | null = null;
-
-  return {
-    dueBy: (rowDate) => {
-      // The counter consumes occurrences as it goes, so a caller that asks out
-      // of order would silently swallow every occurrence between the two dates.
-      // Refuse rather than answer wrongly -- prose could not make this checkable.
-      if (lastRowDate !== null && rowDate < lastRowDate) {
-        throw new Error(
-          `recurringOccurrencesDue: rowDate ${rowDate} precedes ${lastRowDate}; ` +
-            'occurrences are consumed in date order, one call per schedule row',
-        );
-      }
-      lastRowDate = rowDate;
-      let count = 0;
-      for (;;) {
-        if (pendingDueIso === null) {
-          pendingDueIso = isoDay(cursor);
-        }
-        // Not yet due: it stays pending for a later payment.
-        if (pendingDueIso > rowDate) break;
-        // Past the window: nothing further is ever due, and the pending date
-        // stays parked so subsequent rows count nothing.
-        if (extra.endDate && pendingDueIso > extra.endDate) break;
-        count++;
-        cursor = advanceByFrequency(cursor, cadence);
-        pendingDueIso = null;
-      }
-      return count;
-    },
-  };
-}
+import {
+  ScheduleFrequency,
+  advanceDate,
+  getPeriodicRate,
+  getPeriodsPerYear,
+  isoDay,
+  PAYOFF_EPSILON,
+  resolveMaxPayments,
+} from "@/lib/loan-frequency";
+import {
+  OverpaymentMode,
+  recurringOccurrencesDue,
+} from "@/lib/loan-overpayments";
+import {
+  LoanScheduleInput,
+  LoanScheduleResult,
+  ScheduleRow,
+  round2,
+  round4,
+} from "@/lib/loan-schedule-types";
 
 /**
  * Contractual payment for a mortgage amortized over a given period.
@@ -802,18 +64,33 @@ export function calculateMortgagePaymentAmount(
 ): number {
   if (principal <= 0 || amortizationMonths <= 0) return 0;
 
-  if (frequency === 'ACCELERATED_BIWEEKLY' || frequency === 'ACCELERATED_WEEKLY') {
-    const monthlyRate = getPeriodicRate(annualRate, 12, isCanadian, isVariableRate);
+  if (
+    frequency === "ACCELERATED_BIWEEKLY" ||
+    frequency === "ACCELERATED_WEEKLY"
+  ) {
+    const monthlyRate = getPeriodicRate(
+      annualRate,
+      12,
+      isCanadian,
+      isVariableRate,
+    );
     // Accelerated payments derive from the monthly payment, rounded to
     // storage precision first as the backend does
-    const monthlyPayment = round4(solvePayment(principal, monthlyRate, amortizationMonths));
-    const divisor = frequency === 'ACCELERATED_BIWEEKLY' ? 2 : 4;
+    const monthlyPayment = round4(
+      solvePayment(principal, monthlyRate, amortizationMonths),
+    );
+    const divisor = frequency === "ACCELERATED_BIWEEKLY" ? 2 : 4;
     return round4(monthlyPayment / divisor);
   }
 
   const periodsPerYear = getPeriodsPerYear(frequency);
   const totalPayments = Math.round((amortizationMonths * periodsPerYear) / 12);
-  const periodicRate = getPeriodicRate(annualRate, periodsPerYear, isCanadian, isVariableRate);
+  const periodicRate = getPeriodicRate(
+    annualRate,
+    periodsPerYear,
+    isCanadian,
+    isVariableRate,
+  );
   return round4(solvePayment(principal, periodicRate, totalPayments));
 }
 
@@ -839,11 +116,17 @@ export function calculatePaymentForTerm(
     isVariableRate,
   );
   if (periodicRate === 0) return round4(balance / periods);
-  return round4((balance * periodicRate) / (1 - Math.pow(1 + periodicRate, -periods)));
+  return round4(
+    (balance * periodicRate) / (1 - Math.pow(1 + periodicRate, -periods)),
+  );
 }
 
 /** Standard amortization payment: PMT = P * [r(1+r)^n] / [(1+r)^n - 1] */
-function solvePayment(principal: number, periodicRate: number, totalPayments: number): number {
+function solvePayment(
+  principal: number,
+  periodicRate: number,
+  totalPayments: number,
+): number {
   if (totalPayments <= 0) return 0;
   if (periodicRate === 0) {
     return principal / totalPayments;
@@ -865,7 +148,7 @@ function solvePayment(principal: number, periodicRate: number, totalPayments: nu
 export function generateBudgetSchedule(
   input: LoanScheduleInput,
   budget: number,
-  mode: OverpaymentMode = 'LOWER_INSTALLMENT',
+  mode: OverpaymentMode = "LOWER_INSTALLMENT",
   window: { startDate?: string; endDate?: string } = {},
 ): LoanScheduleResult {
   const {
@@ -888,9 +171,19 @@ export function generateBudgetSchedule(
   // contractual term (it steps down as the balance falls); SHORTEN_TERM keeps
   // the contractual installment fixed and the overpayment constant. Either way
   // the total paid is the budget, so the balance/payoff are identical.
-  const lowerInstallment = mode === 'LOWER_INSTALLMENT';
+  const lowerInstallment = mode === "LOWER_INSTALLMENT";
+  // The same quantity `lowerEnd` is in the ordinary generator, and obtained the
+  // same way: a caller that already has the no-overpayment term supplies it
+  // (`lowerEndPeriod`) rather than paying for a second full schedule per
+  // candidate. Deriving it by two routes is how the two entry points drifted
+  // before -- one of them would keep the recursion after a change to the other.
   const contractualPeriods = lowerInstallment
-    ? Math.max(1, generateLoanSchedule({ ...input, overpayments: undefined }).numPayments)
+    ? Math.max(
+        1,
+        input.lowerEndPeriod ??
+          generateLoanSchedule({ ...input, overpayments: undefined })
+            .numPayments,
+      )
     : 0;
 
   const rateChanges = [...(input.rateChanges ?? [])].sort((a, b) =>
@@ -1019,13 +312,15 @@ export function generateBudgetSchedule(
  * applied after the regular payment each period (capped at the remaining
  * balance), shortening the schedule.
  */
-export function generateLoanSchedule(input: LoanScheduleInput): LoanScheduleResult {
+export function generateLoanSchedule(
+  input: LoanScheduleInput,
+): LoanScheduleResult {
   const budget = input.overpayments?.targetMonthlyPayment;
   if (budget && budget > 0) {
     return generateBudgetSchedule(
       input,
       budget,
-      input.overpayments?.targetMonthlyPaymentMode ?? 'LOWER_INSTALLMENT',
+      input.overpayments?.targetMonthlyPaymentMode ?? "LOWER_INSTALLMENT",
       {
         startDate: input.overpayments?.targetMonthlyPaymentStart,
         endDate: input.overpayments?.targetMonthlyPaymentEnd,
@@ -1049,15 +344,17 @@ export function generateLoanSchedule(input: LoanScheduleInput): LoanScheduleResu
 
   // Each overpayment carries its own mode; SHORTEN_TERM is the default for
   // those that omit one.
-  const modeOf = (m?: OverpaymentMode): OverpaymentMode => m ?? 'SHORTEN_TERM';
+  const modeOf = (m?: OverpaymentMode): OverpaymentMode => m ?? "SHORTEN_TERM";
   const hasOverpayments = Boolean(
     overpayments?.recurringExtra || (overpayments?.lumpSums?.length ?? 0) > 0,
   );
   const anyLowerOverpayment =
     hasOverpayments &&
     ((overpayments?.recurringExtra != null &&
-      modeOf(overpayments.recurringExtra.mode) === 'LOWER_INSTALLMENT') ||
-      (overpayments?.lumpSums ?? []).some((l) => modeOf(l.mode) === 'LOWER_INSTALLMENT'));
+      modeOf(overpayments.recurringExtra.mode) === "LOWER_INSTALLMENT") ||
+      (overpayments?.lumpSums ?? []).some(
+        (l) => modeOf(l.mode) === "LOWER_INSTALLMENT",
+      ));
 
   // Two re-levelling ends. `reLevelEveryPeriod` (a passed fixedEndPeriod) holds
   // a variable-rate contractual schedule on its term every period. `lowerEnd`
@@ -1185,14 +482,19 @@ export function generateLoanSchedule(input: LoanScheduleInput): LoanScheduleResu
       const occurrences = recurringOccurrences.dueBy(rowDate);
       if (occurrences > 0) {
         extraPrincipal += recurringExtra.amount * occurrences;
-        if (modeOf(recurringExtra.mode) === 'LOWER_INSTALLMENT') lowerApplied = true;
+        if (modeOf(recurringExtra.mode) === "LOWER_INSTALLMENT")
+          lowerApplied = true;
       }
     }
     // Lump sums land on the first payment on or after their date (sums dated
     // before the first payment attach to row 1)
-    while (lumpSumIndex < lumpSums.length && lumpSums[lumpSumIndex].date <= rowDate) {
+    while (
+      lumpSumIndex < lumpSums.length &&
+      lumpSums[lumpSumIndex].date <= rowDate
+    ) {
       extraPrincipal += lumpSums[lumpSumIndex].amount;
-      if (modeOf(lumpSums[lumpSumIndex].mode) === 'LOWER_INSTALLMENT') lowerApplied = true;
+      if (modeOf(lumpSums[lumpSumIndex].mode) === "LOWER_INSTALLMENT")
+        lowerApplied = true;
       lumpSumIndex++;
     }
     if (extraPrincipal > balance) {
@@ -1213,7 +515,11 @@ export function generateLoanSchedule(input: LoanScheduleInput): LoanScheduleResu
     // only in the period it lands, stepping the payment down while
     // SHORTEN_TERM overpayments leave it unchanged (shortening the term).
     const reLevelEnd =
-      reLevelEveryPeriod !== null ? reLevelEveryPeriod : lowerApplied ? lowerEnd : null;
+      reLevelEveryPeriod !== null
+        ? reLevelEveryPeriod
+        : lowerApplied
+          ? lowerEnd
+          : null;
     if (reLevelEnd !== null) {
       const remaining = reLevelEnd - paymentNumber;
       if (remaining > 0 && balance > PAYOFF_EPSILON) {
@@ -1258,127 +564,3 @@ export function generateLoanSchedule(input: LoanScheduleInput): LoanScheduleResu
   };
 }
 
-/**
- * The annual rate (percentage) in effect on a given date: the latest row with
- * `effectiveDate <= date`, else the earliest row's rate (a date before the
- * first recorded change still amortizes at the origination rate), else the
- * fallback. Shared by the schedule table (per-row historical rate) and
- * `buildRateTimeline`'s starting rate.
- */
-export function effectiveAnnualRateOn(
-  rows: RateTimelineRow[],
-  dateIso: string,
-  fallbackAnnualRate: number,
-): number {
-  const sorted = [...rows].sort((a, b) =>
-    a.effectiveDate.localeCompare(b.effectiveDate),
-  );
-  const atOrBefore = sorted.filter((row) => row.effectiveDate <= dateIso);
-  if (atOrBefore.length > 0) {
-    return atOrBefore[atOrBefore.length - 1].annualRate;
-  }
-  return sorted[0]?.annualRate ?? fallbackAnnualRate;
-}
-
-/**
- * Resolve a persisted rate history into engine inputs for a schedule that
- * starts at `scheduleStartIso`: the rate in effect at the start is the
- * latest row on or before that date (before the earliest row, the earliest
- * row's rate applies; with no rows, the fallback), the payment in effect is
- * the latest non-null payment on or before the start (before the earliest
- * row, that row's payment applies -- the origination installment), and the
- * remaining rows become steps for generateLoanSchedule.
- */
-export function buildRateTimeline(
-  rows: RateTimelineRow[],
-  scheduleStartIso: string,
-  fallbackAnnualRate: number,
-): RateTimeline {
-  const sorted = [...rows].sort((a, b) =>
-    a.effectiveDate.localeCompare(b.effectiveDate),
-  );
-
-  const atOrBefore = sorted.filter(
-    (row) => row.effectiveDate <= scheduleStartIso,
-  );
-  const startingAnnualRate = effectiveAnnualRateOn(
-    rows,
-    scheduleStartIso,
-    fallbackAnnualRate,
-  );
-  // Mirror the rate's "before the earliest row, the earliest row applies"
-  // fallback for the payment: a schedule starting shortly before the first
-  // recorded row (payment_start_date precedes the first installment, which is
-  // where detection dates the initial row) still starts at the origination
-  // installment that row records. Only the earliest row is consulted -- later
-  // rows describe later rate levels and become steps anyway.
-  const startingPaymentAmount =
-    [...atOrBefore].reverse().find((row) => row.newPaymentAmount != null)
-      ?.newPaymentAmount ??
-    (atOrBefore.length === 0 ? sorted[0]?.newPaymentAmount ?? null : null);
-
-  const rateChanges = sorted
-    .filter((row) => row.effectiveDate > scheduleStartIso)
-    .map((row) => ({
-      effectiveDate: row.effectiveDate,
-      annualRate: row.annualRate,
-      paymentAmount: row.newPaymentAmount ?? null,
-    }));
-
-  return { startingAnnualRate, startingPaymentAmount, rateChanges };
-}
-
-export function compareSchedules(
-  baseline: LoanScheduleResult,
-  scenario: LoanScheduleResult,
-): ScenarioComparison {
-  // Every saving is a difference of two lifetime figures, so both sides must
-  // have paid off within the horizon. A truncated schedule's `numPayments` is
-  // the horizon's row count and its `totalInterest` the interest over it, so
-  // subtracting either from a real lifetime figure says nothing about the loan
-  // (a truncated baseline made every scenario look better than it is, and could
-  // even report a negative saving). `monthsSaved` needs the same gate for a
-  // different reason: `monthsBetween` returns 0 for a missing payoff date, which
-  // reads as "the overpayment bought no time" rather than "not known".
-  const comparable = baseline.paidOff && scenario.paidOff;
-  return {
-    baseline,
-    scenario,
-    paymentsSaved: comparable ? baseline.numPayments - scenario.numPayments : null,
-    monthsSaved: comparable
-      ? monthsBetween(scenario.payoffDate, baseline.payoffDate)
-      : null,
-    interestSaved: comparable
-      ? round2(baseline.totalInterest - scenario.totalInterest)
-      : null,
-    // A truncated schedule's `finalPaymentAmount` is the installment at its last
-    // PROJECTED row, not its last payment -- there is no last payment -- so a
-    // drop measured from it is as unknown as the rest.
-    installmentReduction: comparable
-      ? round2(baseline.finalPaymentAmount - scenario.finalPaymentAmount)
-      : null,
-  };
-}
-
-/** Whole months from `fromDate` to `toDate` (0 when either is missing) */
-export function monthsBetween(fromDate: string | null, toDate: string | null): number {
-  if (!fromDate || !toDate) return 0;
-  const from = parseIsoDateParts(fromDate);
-  const to = parseIsoDateParts(toDate);
-  return (to.year - from.year) * 12 + (to.month - from.month);
-}
-
-function parseIsoDateParts(isoDate: string): { year: number; month: number } {
-  const [year, month] = isoDate.split('-').map(Number);
-  return { year, month };
-}
-
-/** Round to 2 decimals (cents), avoiding floating-point drift. */
-export function round2(value: number): number {
-  return Math.round(value * 100) / 100;
-}
-
-/** Storage precision (decimal(20,4)), matching backend roundMoney */
-function round4(value: number): number {
-  return Math.round(value * 10000) / 10000;
-}
